+++ apps/vm-runner/.gitignore
@@ -1,0 +1,1 @@
+run.sh
